chore(e2e): consolidate collection state-store tests (audit D)
Pins the shared per-collection localStorage state stores in one
dedicated spec, replacing the scattered coverage that lived in
three different files. The same store keys (`collection_sorts` and
`collection_view_modes`) are read AND written by both the standalone
`/collections/:slug` route AND the embedded `presentCollection` card;
both halves of that contract now sit next to each other so a
regression on either surface fails the same spec.

Rename + extend:
- collection-sort-persist.spec.ts → collection-state-persist.spec.ts
- Adds two describe blocks: "standalone … shared state stores" (3
  tests) and "embedded presentCollection card … shared state stores"
  (3 tests).

Moved tests:
- "restores the last-used view mode after a reload" — from
  collection-calendar.spec.ts L422-432; the calendar spec keeps its
  view-specific coverage (day view, undated tray, colouring, etc.).
- "embedded card honours the shared localStorage view-mode store" — from
  present-collection.spec.ts; brings the `DATED_EVENTS_DETAIL`
  fixture with it (no other test in present-collection used it).
- "embedded card honours the shared localStorage table sort" — from
  present-collection.spec.ts.
- "sorting in an embedded card writes the shared store" — from
  present-collection.spec.ts.

Kept in present-collection.spec.ts:
- presentCollection card renders without crashing (pageerror sentinel)
- Edit on an open record swaps the modal to the edit form
- Add opens the create form in the shared record modal
- saving an edit returns to the record's detail (does not close)

Numbers:
- e2e mock specs: 65 (unchanged — net rename, no file count delta)
- LOC: ~771 (target file 425, present-collection 154, calendar -14)
  vs ~861 before — net **−90 LOC**, no test() block deletions.
